license-authority/requirements.txt
fastapi
uvicorn[standard]
cryptography
sqlalchemy
pydantic
passlib[bcrypt]
python-jose[cryptography]
